+++ modules/AWSPowerShell/Cmdlets/Route53Resolver/Basic/Get-R53RResolverConfig-Cmdlet.cs
@@ -28,7 +28,7 @@ using Amazon.Route53Resolver.Model;
 namespace Amazon.PowerShell.Cmdlets.R53R
 {
     /// <summary>
-    /// Retrieves the behavior configuration of Route 53 Resolver behavior for a single VPC
+    /// Retrieves the behavior configuration of Route 53 Resolver behavior for a single VPC
     /// from Amazon Virtual Private Cloud.
     /// </summary>
     [Cmdlet("Get", "R53RResolverConfig")]